fix: resolve -Wsign-conversion errors on GCC by adding explicit casts
GCC's -Wconversion includes -Wsign-conversion which caught 5 issues:
- sim_fault.c: return uint32_t as int (rule_count++)
- sim_fault.c: offset int to size_t conversion in export_rules
- test_framework.c: return uint32_t failed_cases as int
- test_framework.c: return uint32_t failed as int (param test)

+++ tests/sim/test_framework.c
@@ -198,7 +198,7 @@ int test_run_suite(test_suite_t* suite, void* ctx) {
         tc = tc->next;
     }
 
-    return g_stats.failed_cases;
+    return (int)g_stats.failed_cases;
 }
 
 int test_run_all(void* ctx) {
@@ -289,7 +289,7 @@ int test_run_parameterized(
         log_output("  [FAIL] %u out of %u parameter sets failed\n", failed, count);
     }
 
-    return failed;
+    return (int)failed;
 }
 
 /* ═══════════════════════════════════════════════════════════════════════════
